Implementing a visitation calculator.
Simple algorithm thats walks the graph of narrower concepts and
calculates the nested sets.

Because we support polyhierarchy, the concept id is nog guaranteed to be
unique. In a nested set that represent a monohierarchy, the formula
max(rght) / 2 represents the number of concepts in that tree. In a
polyhierarchy this would represent the numer of concept positions in
that tree (because a concept can appear in different places in that
tree).