Added a concept_id.
A concept now has a id and a concept_id. ID is the primary key in the
database for the concept table. concept_id is the id of a concept within
a certain conceptscheme.

This means that we can have two concepts with the concept_id 5, but only
one can have the id 5 (and it needn't be one of the two with the
concept_id 5).